+++ miden/tests/integration/operations/io_ops/mem_ops.rs
@@ -1,8 +1,3 @@
-use processor::ContextId;
-use prover::ExecutionError;
-use test_utils::expect_exec_error_matches;
-use vm_core::FieldElement;
-
 use super::{apply_permutation, build_op_test, build_test, Felt, ToElements, TRUNCATE_STACK_PROC};
 
 // LOADING SINGLE ELEMENT ONTO THE STACK (MLOAD)
@@ -237,7 +232,7 @@ fn read_after_write() {
 // MISC
 // ================================================================================================
 
-/// Ensures that the processor returns an error when 2 memory operations occur in the same context,
+/// Ensures that the processor returns no error when 2 memory operations occur in the same context,
 /// at the same address, and in the same clock cycle (which is what RCOMBBASE does when `stack[13] =
 /// stack[14] = 0`).
 #[test]
@@ -246,9 +241,5 @@ fn mem_reads_same_clock_cycle() {
 
     let test = build_test!(asm_op);
 
-    expect_exec_error_matches!(
-        test,
-        ExecutionError::DuplicateMemoryAccess{ctx, addr, clk }
-        if ctx == ContextId::from(0) && addr == 0 && clk == Felt::ONE
-    );
+    test.prove_and_verify(vec![0], false);
 }